@@ -1,5 +1,6 @@
 {
 	"dependencies": {
+		"@aws-sdk/client-s3": "3.437.0",
 		"@hookform/resolvers": "^3.9.0",
 		"@node-rs/argon2": "^2.0.0",
 		"@radix-ui/react-accordion": "^1.2.1",
@@ -28,19 +29,20 @@
 		"date-fns": "^4.1.0",
 		"deepmerge": "^4.3.1",
 		"i18n": "workspace:*",
-		"jotai": "^2.10.1",
+		"jotai": "2.8.0",
 		"js-cookie": "^3.0.5",
 		"logs": "workspace:*",
 		"lucide-react": "^0.453.0",
 		"mail": "workspace:*",
-		"next": "14.2.15",
+		"next": "15.0.1",
 		"next-intl": "3.23.4",
 		"next-themes": "^0.3.0",
 		"nextjs-toploader": "^3.7.15",
 		"nprogress": "^0.2.0",
-		"react": "18.3.1",
+		"oslo": "^1.2.1",
+		"react": "19.0.0-rc-65a56d0e-20241020",
 		"react-cropper": "^2.3.3",
-		"react-dom": "18.3.1",
+		"react-dom": "19.0.0-rc-65a56d0e-20241020",
 		"react-dropzone": "^14.2.10",
 		"react-hook-form": "^7.53.1",
 		"server-only": "^0.0.1",
@@ -66,8 +68,8 @@
 		"@types/mdx": "^2.0.13",
 		"@types/node": "22.7.9",
 		"@types/nprogress": "^0.2.3",
-		"@types/react": "18.3.12",
-		"@types/react-dom": "18.3.1",
+		"@types/react": "npm:types-react@19.0.0-rc.1",
+		"@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
 		"@types/uuid": "^10.0.0",
 		"auth": "workspace:*",
 		"autoprefixer": "10.4.20",
@@ -86,7 +88,7 @@
 	"private": true,
 	"scripts": {
 		"build": "next build",
-		"dev": "next dev",
+		"dev": "next dev --turbo",
 		"e2e": "dotenv -c -e ../../.env -- start-server-and-test dev http://localhost:3000 \"cypress open --e2e\"",
 		"lint": "next lint",
 		"shadcn-ui": "pnpm dlx shadcn@latest",
