@@ -8,7 +8,7 @@
 		"database": "workspace:*",
 		"logs": "workspace:*",
 		"mail": "workspace:*",
-		"next": "14.2.15",
+		"next": "15.0.1",
 		"openai": "^4.68.4",
 		"storage": "workspace:*",
 		"stripe": "^17.2.1",
