@@ -7,7 +7,7 @@
 		"cookie": "^1.0.1",
 		"database": "workspace:*",
 		"logs": "workspace:*",
-		"next": "14.2.15",
+		"next": "15.0.1",
 		"utils": "workspace:*",
 		"zod": "^3.23.8"
 	},
