@@ -9,7 +9,8 @@ import { defineAbilitiesFor } from "../modules/auth/abilities";
 export async function createContext(
 	params?: FetchCreateContextFnOptions | { isAdmin?: boolean },
 ) {
-	const sessionId = cookies().get(config.auth.sessionCookieName)?.value ?? null;
+	const sessionId =
+		(await cookies()).get(config.auth.sessionCookieName)?.value ?? null;
 	const { user, session } = sessionId
 		? await validateSessionToken(sessionId)
 		: { user: null, session: null };
@@ -45,7 +46,7 @@ export async function createContext(
 		teamMemberships,
 	});
 
-	const locale = (cookies().get(config.i18n.localeCookieName)?.value ??
+	const locale = ((await cookies()).get(config.i18n.localeCookieName)?.value ??
 		config.i18n.defaultLocale) as Locale;
 
 	return {
